MKB-671 Carousel move to modules
