+++ next/components/HomePage/SectionLibraryNews.tsx
@@ -3,7 +3,7 @@ import { ComponentHomepageNewsSection, NoticeListingEntityFragment } from '@serv
 import { useTranslation } from 'next-i18next'
 import React from 'react'
 
-import Carousel from '../Molecules/Carousel/Carousel'
+import Carousel from '../../modules/common/Carousel/Carousel'
 import ListingCard from '../Molecules/ListingCard'
 
 interface LibraryNewsProps {
