+++ next/modules/common/Carousel/SliderControl/SliderControl.tsx
